removing all .delete message

+++ src/moona-bibilothecary/commands/AddCommand.ts
@@ -10,18 +10,18 @@ import { Command, CommandInterface } from './commandInterface';
 import { embedAdder } from './embedCreator';
 
 export class AddCommand extends Command implements CommandInterface {
-  commandNames = ["add", "a"];
+  commandNames = ['add', 'a'];
   args = [
     {
-      name: "line No.",
-      type: "number | number.number - (Where the latter number is ss_ eg. song 10 (ss 2))",
-      description: "Number of line that you want to edit",
+      name: 'line No.',
+      type: 'number | number.number - (Where the latter number is ss_ eg. song 10 (ss 2))',
+      description: 'Number of line that you want to edit',
     },
   ];
   timeLimit = config.timeLimit;
 
   help(): string {
-    return this.helpText("Insert a line from Moona-Librarian Post at line");
+    return this.helpText('Insert a line from Moona-Librarian Post at line');
   }
 
   async getSongSeasonNumber(originalMessage: Message, args: Array<string>) {
@@ -29,36 +29,28 @@ export class AddCommand extends Command implements CommandInterface {
     const [, ...others] = args;
     const songNumberText = others.join();
     if (!ID) throw "Please specify post's ID.";
-    if (!songNumberText) throw "Please specify song number.";
+    if (!songNumberText) throw 'Please specify song number.';
     const res1 = /^\s*(\d+)(?:\.(\d+))?\s*$/.exec(songNumberText);
     if (res1) {
       // Pattern of song numbers
       const [, songNumberStr, seasonNumberStr] = res1;
       const songNumber = Number(songNumberStr);
       let seasonNumber: number | undefined;
-      if (!seasonNumberStr || Number(seasonNumberStr) === 1)
-        seasonNumber = undefined;
+      if (!seasonNumberStr || Number(seasonNumberStr) === 1) seasonNumber = undefined;
       else seasonNumber = Number(seasonNumberStr);
-      if (Number.isNaN(songNumber) || songNumber <= 0)
-        throw "Invalid song number.";
+      if (Number.isNaN(songNumber) || songNumber <= 0) throw 'Invalid song number.';
       if (seasonNumber && (Number.isNaN(seasonNumber) || seasonNumber <= 0))
-        throw "Invalid season number.";
+        throw 'Invalid season number.';
       if (!songNumber)
         throw `Invalid song number format\n\tplease use "1" for 1st song or "1.2" for 1st song (ss2)`;
-      if (originalMessage.author.id !== client.application?.id)
-        throw `I can only edit my post.`;
-      if (!originalMessage.embeds.length)
-        throw "Invalid post: no embed detected";
-      return { songNumber, seasonNumber, type: "full" };
+      if (originalMessage.author.id !== client.application?.id) throw `I can only edit my post.`;
+      if (!originalMessage.embeds.length) throw 'Invalid post: no embed detected';
+      return { songNumber, seasonNumber, type: 'full' };
     }
     throw `Error parsing ${songNumberText}`;
   }
 
-  findLowerBound(
-    lines: Array<LineComponent>,
-    songNumber: number,
-    seasonNumber = 1
-  ) {
+  findLowerBound(lines: Array<LineComponent>, songNumber: number, seasonNumber = 1) {
     for (let i = 0; i < lines.length; i++) {
       if (songNumber === lines[i].songNumber) {
         const lineSeasonNumber = lines[i].seasonNumber || 1;
@@ -68,11 +60,7 @@ export class AddCommand extends Command implements CommandInterface {
     return lines.length;
   }
 
-  async getEmbedEditor(
-    originalMessage: Message,
-    songNumber: number,
-    seasonNumber?: number
-  ) {
+  async getEmbedEditor(originalMessage: Message, songNumber: number, seasonNumber?: number) {
     const { description, url } = originalMessage.embeds[0];
     const lines: Array<LineComponent> = [];
     description?.split(/\n/g).forEach((line) => {
@@ -83,9 +71,7 @@ export class AddCommand extends Command implements CommandInterface {
 
     const foundIdx = this.findLowerBound(lines, songNumber, seasonNumber);
     const found = lines[foundIdx] ? lines[foundIdx] : new LineComponent();
-    const beforeFound = lines[foundIdx - 1]
-      ? lines[foundIdx - 1]
-      : new LineComponent();
+    const beforeFound = lines[foundIdx - 1] ? lines[foundIdx - 1] : new LineComponent();
 
     return { lines, foundIdx, beforeFound, found, url: url || undefined };
   }
@@ -95,8 +81,7 @@ export class AddCommand extends Command implements CommandInterface {
     let currentMessage = indexMesssage;
     while (i > 0 && currentMessage.reference?.messageId) {
       const parentMessage = await currentMessage.fetchReference();
-      if (currentMessage.deletable && !currentMessage.hasThread)
-        await currentMessage.delete();
+      if (currentMessage.deletable && !currentMessage.hasThread) await currentMessage.delete();
       currentMessage = parentMessage;
       i--;
     }
@@ -105,7 +90,7 @@ export class AddCommand extends Command implements CommandInterface {
   awaitReply(
     message: Message,
     beforeLine: LineComponent,
-    line: LineComponent
+    line: LineComponent,
   ): Promise<Message | undefined> {
     return new Promise(async (resolve, reject) => {
       try {
@@ -113,55 +98,53 @@ export class AddCommand extends Command implements CommandInterface {
           `Reply to this message with the message to edit, or press ${
             moonaEmoji.denied
           } to cancel\n> Pattern: "New line content" | "New line content" "timestamp" - (Where "timestamp" is \\_h\\_m\\_s)\n\`\`\`${beforeLine.toDisplay()}\n> ...\n${line.toDisplay()}\`\`\`(This message will be deleted automatically <t:${getUnixTime(
-            addMilliseconds(new Date(), this.timeLimit)
-          )}:R>)`
+            addMilliseconds(new Date(), this.timeLimit),
+          )}:R>)`,
         );
 
         // Reaction Collector
         const reactionCollector = messageReply.createReactionCollector({
           filter: (reaction, user) =>
-            reaction.emoji.name === moonaEmoji.string.denied &&
-            user.id === message.author.id,
+            reaction.emoji.name === moonaEmoji.string.denied && user.id === message.author.id,
           time: this.timeLimit,
         });
         // Message Collector
         const messageCollector = messageReply.channel.createMessageCollector({
-          filter: (messageAnswerReply) =>
-            messageAnswerReply.author.id === message.author.id,
+          filter: (messageAnswerReply) => messageAnswerReply.author.id === message.author.id,
           time: this.timeLimit,
         });
 
-        reactionCollector.on("collect", () => {
-          reactionCollector.stop("cancel");
+        reactionCollector.on('collect', () => {
+          reactionCollector.stop('cancel');
         });
-        reactionCollector.on("end", async (collected, reason) => {
-          if (reason === "accept") {
+        reactionCollector.on('end', async (collected, reason) => {
+          if (reason === 'accept') {
             const [, , ...newMsg] = messageReply.content.split(/\n/g);
             const [lastMsg] = newMsg[newMsg.length - 1].split(
-              "(This message will be deleted automatically"
+              '(This message will be deleted automatically',
             );
-            const finalMsg = [...newMsg.slice(0, -1), lastMsg].join("\n");
+            const finalMsg = [...newMsg.slice(0, -1), lastMsg].join('\n');
             if (messageReply.editable) messageReply.edit(finalMsg);
             await messageReply.reactions.removeAll();
           } else {
             this.disposeMessage(messageReply, 2);
-            messageCollector.stop("cancel");
+            messageCollector.stop('cancel');
             resolve(undefined);
           }
         });
 
-        messageCollector.on("collect", (messageAnswerReply) => {
+        messageCollector.on('collect', (messageAnswerReply) => {
           if (messageAnswerReply.reference?.messageId === messageReply.id) {
             // If there is a reply to current message
             resolve(messageAnswerReply);
-            messageCollector.stop("accept");
-            reactionCollector.stop("accept");
+            messageCollector.stop('accept');
+            reactionCollector.stop('accept');
           } else {
             // There is no reply, must be an accident, proceed to delete the message
             if (messageAnswerReply.deletable) messageAnswerReply.delete();
           }
         });
-        if (!messageReply.deleted) await messageReply.react(moonaEmoji.denied);
+        await messageReply.react(moonaEmoji.denied);
       } catch (err) {
         reject(err);
       }
@@ -175,17 +158,13 @@ export class AddCommand extends Command implements CommandInterface {
       url?: string;
       songNumber: number;
       seasonNumber?: number;
-    }
-  ): Promise<
-    | { type: "accept"; newLine: LineComponent }
-    | { type: "cancel" }
-    | { type: "edit" }
-  > {
+    },
+  ): Promise<{ type: 'accept'; newLine: LineComponent } | { type: 'cancel' } | { type: 'edit' }> {
     return new Promise(async (resolve, reject) => {
       try {
         const [, addLine, addTimestamp] =
           /^(?:\d+(?:\s*\(\s*\d+\s*\)\s*)?[.\s]*)?(.*?)\s*((?:\d+h)?(?:\d+m)?\d+s)?\s*$/.exec(
-            messageAnswerReply.content
+            messageAnswerReply.content,
           ) || [];
         const { url, songNumber, seasonNumber } = newLineObj;
 
@@ -196,16 +175,14 @@ export class AddCommand extends Command implements CommandInterface {
             seasonNumber,
             timestamp: addTimestamp,
           },
-          url
+          url,
         );
 
         const finalMessage = await messageAnswerReply.reply({
-          content: `Press ${moonaEmoji.approb} for confirmation, ${
-            moonaEmoji.denied
-          } for cancel, ${
+          content: `Press ${moonaEmoji.approb} for confirmation, ${moonaEmoji.denied} for cancel, ${
             moonaEmoji.edit
           } for editing again\n(This message will be deleted automatically <t:${getUnixTime(
-            addMilliseconds(new Date(), this.timeLimit)
+            addMilliseconds(new Date(), this.timeLimit),
           )}:R>)`,
           embeds: [
             embedAdder({
@@ -221,36 +198,33 @@ export class AddCommand extends Command implements CommandInterface {
         // Reaction Collector
         const reactionCollector = finalMessage.createReactionCollector({
           filter: (reaction, user) =>
-            [
-              moonaEmoji.string.approb,
-              moonaEmoji.string.edit,
-              moonaEmoji.string.denied,
-            ].includes(reaction.emoji.name || "") &&
-            user.id === messageAnswerReply.author.id,
+            [moonaEmoji.string.approb, moonaEmoji.string.edit, moonaEmoji.string.denied].includes(
+              reaction.emoji.name || '',
+            ) && user.id === messageAnswerReply.author.id,
           time: this.timeLimit,
         });
-        reactionCollector.on("collect", (collected) => {
-          reactionCollector.stop(collected.emoji.name || "");
+        reactionCollector.on('collect', (collected) => {
+          reactionCollector.stop(collected.emoji.name || '');
         });
-        reactionCollector.on("end", async (collected, reason) => {
+        reactionCollector.on('end', async (collected, reason) => {
           switch (reason) {
             case moonaEmoji.string.approb:
               this.disposeMessage(finalMessage, 4);
-              resolve({ type: "accept", newLine });
+              resolve({ type: 'accept', newLine });
               break;
             case moonaEmoji.string.edit:
               this.disposeMessage(finalMessage, 3);
-              resolve({ type: "edit" });
+              resolve({ type: 'edit' });
               break;
             case moonaEmoji.string.denied:
             default:
               this.disposeMessage(finalMessage, 4);
-              resolve({ type: "cancel" });
+              resolve({ type: 'cancel' });
           }
         });
-        if (!finalMessage.deleted) await finalMessage.react(moonaEmoji.approb);
-        if (!finalMessage.deleted) await finalMessage.react(moonaEmoji.denied);
-        if (!finalMessage.deleted) await finalMessage.react(moonaEmoji.edit);
+        await finalMessage.react(moonaEmoji.approb);
+        await finalMessage.react(moonaEmoji.denied);
+        await finalMessage.react(moonaEmoji.edit);
       } catch (err) {
         reject(err);
       }
@@ -261,7 +235,7 @@ export class AddCommand extends Command implements CommandInterface {
     lines: Array<LineComponent>,
     foundIdx: number,
     songNumber: number,
-    seasonNumber?: number
+    seasonNumber?: number,
   ) {
     if (seasonNumber) {
       // push by season
@@ -301,17 +275,12 @@ export class AddCommand extends Command implements CommandInterface {
   }) {
     const { foundIdx, lines, originalMessage, newLine } = payload;
 
-    this.pushSongIndex(
-      lines,
-      foundIdx,
-      newLine.songNumber,
-      newLine.seasonNumber
-    );
+    this.pushSongIndex(lines, foundIdx, newLine.songNumber, newLine.seasonNumber);
     lines.splice(foundIdx, 0, newLine);
 
     const originalEmbed = originalMessage.embeds[0];
 
-    const newDescription = lines.map((line) => line.toRaw()).join("\n");
+    const newDescription = lines.map((line) => line.toRaw()).join('\n');
 
     originalMessage.edit({
       embeds: [originalEmbed.setDescription(newDescription)],
@@ -325,44 +294,37 @@ export class AddCommand extends Command implements CommandInterface {
         message.guildId === message.reference.guildId &&
         message.reference.messageId
       ) {
-        const originalMessage = await message.channel.messages.fetch(
-          message.reference.messageId
-        );
+        const originalMessage = await message.channel.messages.fetch(message.reference.messageId);
         // Only response to this bot's own post reply.
         if (originalMessage.author.id === client.application?.id) {
           // Find song number from args
           const { songNumber, seasonNumber } = await this.getSongSeasonNumber(
             originalMessage,
-            args
+            args,
           );
           // use args to get line & embed to send into channel
-          const { lines, foundIdx, found, beforeFound, url } =
-            await this.getEmbedEditor(
-              originalMessage,
-              songNumber,
-              seasonNumber
-            );
+          const { lines, foundIdx, found, beforeFound, url } = await this.getEmbedEditor(
+            originalMessage,
+            songNumber,
+            seasonNumber,
+          );
 
-          let userChoice: "accept" | "cancel" | "edit";
+          let userChoice: 'accept' | 'cancel' | 'edit';
           do {
             // send message in channel
             // await message.reply({ embeds });
             // await reply from message
-            const messageAnswerReply = await this.awaitReply(
-              message,
-              beforeFound,
-              found
-            );
+            const messageAnswerReply = await this.awaitReply(message, beforeFound, found);
             // User answered the edited answer, display the confirmation message
 
             if (messageAnswerReply) {
               const confirmedResponse = await this.showConfirmation(
                 messageAnswerReply,
                 originalMessage,
-                { url, songNumber, seasonNumber }
+                { url, songNumber, seasonNumber },
               );
               userChoice = confirmedResponse.type;
-              if (confirmedResponse.type === "accept") {
+              if (confirmedResponse.type === 'accept') {
                 await this.addOriginalEmbed({
                   originalMessage,
                   lines,
@@ -370,8 +332,8 @@ export class AddCommand extends Command implements CommandInterface {
                   newLine: confirmedResponse.newLine,
                 });
               }
-            } else userChoice = "cancel";
-          } while (userChoice === "edit");
+            } else userChoice = 'cancel';
+          } while (userChoice === 'edit');
         }
       }
     }
